requirements.txt
asyncio-nats-client==0.11.5
certifi==2021.10.8
charset-normalizer==2.0.12
configparser==5.2.0
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
mergedeep==1.3.4
pytz==2021.3
requests==2.27.1
urllib3==1.26.8
uuid==1.30
xmltodict==0.12.0
zipp==3.7.0
python-dateutil